% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTime.R
\name{addMetadataToHeader}
\alias{addMetadataToHeader}
\title{Add sample metadata entries to the VCF header}
\usage{
addMetadataToHeader(vcf, purity = NULL, ploidy = NULL, sex = NULL,
  is_wgd = NULL, subclonal_architecture = NULL, cluster_power = NULL)
}
\arguments{
\item{vcf}{A VCF object}

\item{purity}{Sample purity (Default: NULL)}

\item{ploidy}{Sample ploidy (Default: NULL)}

\item{sex}{Donor sex (Default: NULL)}

\item{is_wgd}{Sample whole genome doubling status (TRUE if a whole genome doubling has occurred) (Default: NULL)}

\item{subclonal_architecture}{Sample subclonal architecture data.frame (Default: NULL)}

\item{cluster_power}{Sample vector with a power estimate per mutation cluster (Default: NULL)}
}
\value{
The VCF object with requested entries added to the metadata
}
\description{
Add sample metadata entries to the VCF header
}
\author{
sd11
}
