% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTime.R
\name{addMutTime}
\alias{addMutTime}
\title{Add mutation timing information to vcf object}
\usage{
addMutTime(vcf, V)
}
\arguments{
\item{vcf}{The original `VCF`}

\item{V}{The `DataFrame` with timing information from mutationTime}
}
\value{
The combined vcf with updated header.
}
\description{
Add mutation timing information to vcf object
}
\author{
Moritz Gerstung
}
