% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTime.R
\name{bbToTime}
\alias{bbToTime}
\title{Convert timing parameters into timing estimates}
\usage{
bbToTime(bb, timing_param = bb$timing_param, n.snv_mnv = bb$n.snv_mnv,
  pseudo.count = 5)
}
\arguments{
\item{bb}{}

\item{timing_param}{}

\item{pseudo.count}{}
}
\value{
data.frame()
}
\description{
Convert timing parameters into timing estimates
}
\author{
mg14
}
