% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTime.R
\name{classWgd}
\alias{classWgd}
\title{Heuristic to classify WGD based on CN profile}
\usage{
classWgd(cn)
}
\arguments{
\item{cn}{Copy number with consensus annotation meta data.}
}
\value{
FALSE/TRUE
}
\description{
The heuristic calculates the genome-wide average ploidy (ie total copy number) and also average homozygousity. Whole genome duplicated tumours
were found to have a ploidy in excess of 2.9 - 2 * homozygousity.
}
\author{
mg14
}
