% Generated by roxygen2 (4.0.1): do not edit by hand
\name{defineMcnStates}
\alias{defineMcnStates}
\title{Compute possible Mutation copy number states}
\usage{
defineMcnStates(bb, clusters, purity, gender = "female", isWgd = FALSE)
}
\arguments{
\item{bb}{Copy number with consensus annotation meta data}

\item{clusters}{}

\item{purity}{}

\item{gender}{}

\item{isWgd}{}
}
\value{
list of length nrow(bb), can be added to mcols(bb)
}
\description{
Compute possible Mutation copy number states
}
\author{
mg14
}

