% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTime.R
\name{getClusterPower}
\alias{getClusterPower}
\title{Get mutation cluster power from VCF metadata}
\usage{
getClusterPower(vcf)
}
\arguments{
\item{VCF}{object to query}
}
\value{
The mutation cluster power (vector with an entry per cluster), or NULL if the metadata entry is not available
}
\description{
Get mutation cluster power from VCF metadata
}
\author{
sd11
}
