% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTime.R
\name{getHomozygousity}
\alias{getHomozygousity}
\title{Extract the average homozygousity from copy number}
\usage{
getHomozygousity(cn)
}
\arguments{
\item{cn}{The `GRanges` copy number data}
}
\value{
The numeric value of average homozygousity across the genome.
}
\description{
Extract the average homozygousity from copy number
}
\author{
Moritz Gerstung
}
