% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTime.R
\name{getPloidy}
\alias{getPloidy}
\title{Get sample ploidy from VCF metadata}
\usage{
getPloidy(vcf)
}
\arguments{
\item{VCF}{object to query}
}
\value{
The sample ploidy, or NULL if the metadata entry is not available
}
\description{
Get sample ploidy from VCF metadata
}
\author{
sd11
}
