% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTime.R
\name{getPurity}
\alias{getPurity}
\title{Get sample purity from VCF metadata}
\usage{
getPurity(vcf)
}
\arguments{
\item{VCF}{object to query}
}
\value{
The sample purity, or NULL if the metadata entry is not available
}
\description{
Get sample purity from VCF metadata
}
\author{
sd11
}
