% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTime.R
\name{getSex}
\alias{getSex}
\title{Get donor sex from VCF metadata}
\usage{
getSex(vcf)
}
\arguments{
\item{VCF}{object to query}
}
\value{
The donor sex or, or NULL if the metadata entry is not available
}
\description{
Get donor sex from VCF metadata
}
\author{
sd11
}
