% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTime.R
\name{getSubclonalArchitecture}
\alias{getSubclonalArchitecture}
\title{Get sample subclonal architecture from VCF metadata}
\usage{
getSubclonalArchitecture(vcf)
}
\arguments{
\item{VCF}{object to query}
}
\value{
A data.frame with the subclonal architecture of the sample, or NULL if the metadata entry is not available
}
\description{
Get sample subclonal architecture from VCF metadata
}
\author{
sd11
}
