% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTime.R
\name{getWGDstatus}
\alias{getWGDstatus}
\title{Get sample whole genome doubling status from VCF metadata}
\usage{
getWGDstatus(vcf)
}
\arguments{
\item{VCF}{object to query}
}
\value{
The sample whole genome doubling status (TRUE if a WGD occurred, FALSE otherwise), or NULL if the metadata entry is not available
}
\description{
Get sample whole genome doubling status from VCF metadata
}
\author{
sd11
}
