% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTime.R
\name{mtToTime}
\alias{mtToTime}
\title{Convert timing parameters into timing estimates}
\usage{
mtToTime(
  cn,
  timing_param = cn$timing_param,
  n.snv_mnv = cn$n.snv_mnv,
  pseudo.count = 5
)
}
\arguments{
\item{cn}{Copy number input}

\item{timing_param}{}

\item{pseudo.count}{}
}
\value{
data.frame()
}
\description{
Convert timing parameters into timing estimates
}
\author{
mg14
}
