% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTime.R
\name{plotSample}
\alias{plotSample}
\title{Plot timing}
\usage{
plotSample(
  vcf,
  cn,
  sv = NULL,
  title = "",
  regions = NULL,
  ylim.cn = c(0, 5),
  layout.height = c(4, 1.2, 3.5),
  y.sv = ylim.cn[2] - 1
)
}
\arguments{
\item{vcf}{The `VCF` file to plot, with timing information added}

\item{cn}{The copy number `GRanges` object}

\item{sv}{Optional structural variants information.}

\item{title}{The titel of the plot. Default = "".}

\item{regions}{The region to show as a `GRanges` object. Default is to show all chromosomes.}

\item{ylim.cn}{The y-limits for the copy number stacked bar plot.}

\item{layout.height}{The relative proportions of the layout.}

\item{y.sv}{Where to plot arches for SVs.}
}
\description{
Plot timing
}
\author{
mg14
}
