% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTime.R
\name{readVcf}
\alias{readVcf}
\title{Read a VCF file}
\usage{
readVcf(file, ...)
}
\arguments{
\item{file}{Path to a VCF file}
}
\value{
A VCF object
}
\description{
This function overloads VariantAnnotation::readVcf as the pAllSubclones info column requires additional encoding to fit within the VCF standard
}
\author{
sd11
}
