% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTime.R
\name{writeVcf}
\alias{writeVcf}
\title{Write a VCF file}
\usage{
writeVcf(vcf, ...)
}
\arguments{
\item{vcf}{A VCF object}
}
\description{
This function overloads VariantAnnotation::writeVcf as the pAllSubclones info column requires additional encoding to fit within the VCF standard
}
\author{
sd11
}
