% Generated by roxygen2 (4.0.1): do not edit by hand
\name{classifyMutations}
\alias{classifyMutations}
\title{Classify Mutations}
\usage{
classifyMutations(x, reclassify = c("missing", "all", "none"))
}
\arguments{
\item{x}{}

\item{reclassify}{}
}
\value{
factor()
}
\description{
Classify Mutations
}
\author{
mg14
}

