% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationTime.R
\name{plotSample}
\alias{plotSample}
\title{Plot timing}
\usage{
plotSample(vcf, bb, sv = NA, title = "", regions = NULL, ylim.bb = c(0,
  5), layout.height = c(4, 1.2, 3.5), y1 = ylim.bb[2] - 1)
}
\arguments{
\item{vcf}{}

\item{bb}{}

\item{sv}{}

\item{title}{}

\item{regions}{}

\item{ylim.bb}{}

\item{layout.height}{}

\item{y1}{}
}
\description{
Plot timing
}
\author{
mg14
}
